requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
passlib[argon2]
python-jose[cryptography]
pydantic-settings
email-validator
asyncpg
greenlet
argon2-cffi
psycopg2-binary
